+++ cmd/state-installer/internal/installer/installation_darwin.go
@@ -7,19 +7,22 @@ import (
 	"path/filepath"
 
 	"github.com/ActiveState/cli/internal/appinfo"
+	"github.com/ActiveState/cli/internal/constants"
 	"github.com/ActiveState/cli/internal/errs"
 	"github.com/ActiveState/cli/internal/fileutils"
 )
 
-const (
-	appName = "ActiveState Desktop.app"
-)
-
 // InstallSystemFiles installs files in the /Application directory
 func InstallSystemFiles(fromDir, binaryDir, systemInstallPath string) error {
-	err := os.RemoveAll(filepath.Join(systemInstallPath, appName))
+	err := os.RemoveAll(filepath.Join(systemInstallPath, constants.MacOSApplicationName))
 	if err != nil {
 		return errs.Wrap(err, "Could not remove old app directory")
+	}
+
+	// ensure systemInstallPath exists
+	err = fileutils.MkdirUnlessExists(systemInstallPath)
+	if err != nil {
+		return errs.Wrap(err, "Application directory %s did not exist, and failed to create it", systemInstallPath)
 	}
 
 	err = fileutils.MoveAllFilesCrossDisk(fromDir, systemInstallPath)
@@ -28,7 +31,7 @@ func InstallSystemFiles(fromDir, binaryDir, systemInstallPath string) error {
 	}
 
 	fromTray := appinfo.TrayApp(binaryDir)
-	toTray := appinfo.TrayApp(filepath.Join(systemInstallPath, appName, "Contents", "MacOS"))
+	toTray := appinfo.TrayApp(filepath.Join(systemInstallPath, constants.MacOSApplicationName, "Contents", "MacOS"))
 	err = createNewSymlink(fromTray.Exec(), toTray.Exec())
 	if err != nil {
 		return errs.Wrap(err, "Could not create state-tray symlink")