+++ internal/runners/clean/config.go
@@ -14,6 +14,7 @@ type configurable interface {
 	project.ConfigAble
 	ConfigPath() string
 	CachePath() string
+	GetInt(string) int
 }
 
 type Config struct {
@@ -54,5 +55,5 @@ func (c *Config) Run(params *ConfigParams) error {
 	}
 
 	logging.Debug("Removing config directory: %s", c.cfg.ConfigPath())
-	return removeConfig(c.cfg)
+	return removeConfig(c.cfg.ConfigPath(), c.output)
 }